requirements.txt
python-dotenv==1.0.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2==2.9.9


